ai-ml-backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
python-multipart==0.0.9
google-generativeai==0.3.2
langchain==0.1.9
langchain-google-genai==0.0.11
pypdf==4.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
pytest==8.0.1
httpx==0.26.0
python-docx==1.0.1
PyMuPDF==1.23.7
pdfplumber==0.10.3
python-jose[cryptography]==3.3.0
python-docx==1.0.1
PyMuPDF==1.23.7
pdfplumber==0.10.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0 